requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
typer[all]>=0.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.0.0
pydantic-settings>=2.0.0
